Rename calibration package
