+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/curve/CalibrationDerivative.java
@@ -3,24 +3,26 @@
  * 
  * Please see distribution for license.
  */
-package com.opengamma.strata.pricer.calibration;
+package com.opengamma.strata.pricer.curve;
 
 import java.util.List;
 import java.util.function.Function;
 
 import com.opengamma.strata.collect.array.DoubleArray;
+import com.opengamma.strata.collect.array.DoubleMatrix;
+import com.opengamma.strata.market.curve.CurveParameterSize;
 import com.opengamma.strata.pricer.rate.ImmutableRatesProvider;
 import com.opengamma.strata.product.ResolvedTrade;
 
 /**
- * Provides the calibration value.
+ * Provides the calibration derivative.
  * <p>
- * This provides the value from the specified {@link CalibrationMeasures} instance
- * in matrix form suitable for use in curve calibration root finding.
+ * This provides the value sensitivity from the specified {@link CalibrationMeasures}
+ * instance in matrix form suitable for use in curve calibration root finding.
  * The value will typically be par spread or converted present value.
  */
-class CalibrationValue
-    implements Function<DoubleArray, DoubleArray> {
+class CalibrationDerivative
+    implements Function<DoubleArray, DoubleMatrix> {
 
   /**
    * The trades.
@@ -34,6 +36,11 @@ class CalibrationValue
    * The provider generator, used to create child providers.
    */
   private final RatesProviderGenerator providerGenerator;
+  /**
+   * Provide the order in which the curves appear in the long vector result.
+   * The expected number of parameters for each curve is also provided.
+   */
+  private final List<CurveParameterSize> curveOrder;
 
   /**
    * Creates an instance.
@@ -41,24 +48,28 @@ class CalibrationValue
    * @param trades  the trades
    * @param measures  the calibration measures
    * @param providerGenerator  the provider generator, used to create child providers
+   * @param curveOrder  the curve order
    */
-  CalibrationValue(
+  public CalibrationDerivative(
       List<ResolvedTrade> trades,
       CalibrationMeasures measures,
-      RatesProviderGenerator providerGenerator) {
+      RatesProviderGenerator providerGenerator,
+      List<CurveParameterSize> curveOrder) {
 
-    this.trades = trades;
     this.measures = measures;
+    this.trades = trades;
     this.providerGenerator = providerGenerator;
+    this.curveOrder = curveOrder;
   }
 
   //-------------------------------------------------------------------------
   @Override
-  public DoubleArray apply(DoubleArray x) {
+  public DoubleMatrix apply(DoubleArray x) {
     // create child provider from matrix
-    ImmutableRatesProvider childProvider = providerGenerator.generate(x);
-    // calculate value for each trade using the child provider
-    return DoubleArray.of(trades.size(), i -> measures.value(trades.get(i), childProvider));
+    ImmutableRatesProvider provider = providerGenerator.generate(x);
+    // calculate derivative for each trade using the child provider
+    int size = trades.size();
+    return DoubleMatrix.ofArrayObjects(size, size, i -> measures.derivative(trades.get(i), provider, curveOrder));
   }
 
 }
