+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/curve/CalibrationMeasures.java
@@ -3,7 +3,7 @@
  * 
  * Please see distribution for license.
  */
-package com.opengamma.strata.pricer.calibration;
+package com.opengamma.strata.pricer.curve;
 
 import static com.opengamma.strata.collect.Guavate.toImmutableMap;
