+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/curve/SyntheticCurveCalibrator.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.pricer.calibration;
+package com.opengamma.strata.pricer.curve;
 
 import java.time.LocalDate;
 import java.util.HashMap;
