+++ modules/pricer/src/test/java/com/opengamma/strata/pricer/curve/CalibrationDiscountingSimpleEur3Test.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.pricer.calibration;
+package com.opengamma.strata.pricer.curve;
 
 import static com.opengamma.strata.basics.currency.Currency.EUR;
 import static org.testng.Assert.assertEquals;
