+++ modules/pricer/src/test/java/com/opengamma/strata/pricer/curve/CalibrationZeroRateAndDiscountFactorUsd2OisIrsTest.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.pricer.calibration;
+package com.opengamma.strata.pricer.curve;
 
 import static com.opengamma.strata.basics.currency.Currency.USD;
 import static com.opengamma.strata.basics.date.DayCounts.ACT_365F;
