+++ modules/pricer/src/test/java/com/opengamma/strata/pricer/curve/CalibrationZeroRateUsd2OisFuturesIrsTest.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.pricer.calibration;
+package com.opengamma.strata.pricer.curve;
 
 import static com.opengamma.strata.basics.currency.Currency.USD;
 import static com.opengamma.strata.basics.date.BusinessDayConventions.FOLLOWING;
@@ -11,11 +11,9 @@ import static com.opengamma.strata.basics.date.DayCounts.ACT_360;
 import static com.opengamma.strata.basics.date.DayCounts.ACT_365F;
 import static com.opengamma.strata.basics.date.HolidayCalendarIds.USNY;
 import static com.opengamma.strata.basics.index.IborIndices.USD_LIBOR_3M;
-import static com.opengamma.strata.basics.index.IborIndices.USD_LIBOR_6M;
 import static com.opengamma.strata.basics.index.OvernightIndices.USD_FED_FUND;
 import static com.opengamma.strata.product.swap.type.FixedIborSwapConventions.USD_FIXED_6M_LIBOR_3M;
 import static com.opengamma.strata.product.swap.type.FixedOvernightSwapConventions.USD_FIXED_1Y_FED_FUND_OIS;
-import static com.opengamma.strata.product.swap.type.IborIborSwapConventions.USD_LIBOR_3M_LIBOR_6M;
 import static org.testng.Assert.assertEquals;
 
 import java.time.LocalDate;
@@ -30,13 +28,13 @@ import java.util.function.Function;
 
 import org.testng.annotations.Test;
 
-import com.google.common.collect.ImmutableList;
 import com.opengamma.strata.basics.ReferenceData;
 import com.opengamma.strata.basics.StandardId;
 import com.opengamma.strata.basics.currency.Currency;
 import com.opengamma.strata.basics.currency.CurrencyAmount;
 import com.opengamma.strata.basics.currency.MultiCurrencyAmount;
 import com.opengamma.strata.basics.date.BusinessDayAdjustment;
+import com.opengamma.strata.basics.date.DateSequences;
 import com.opengamma.strata.basics.date.DayCount;
 import com.opengamma.strata.basics.date.DaysAdjustment;
 import com.opengamma.strata.basics.date.Tenor;
@@ -62,15 +60,13 @@ import com.opengamma.strata.market.observable.QuoteId;
 import com.opengamma.strata.market.param.CurrencyParameterSensitivities;
 import com.opengamma.strata.market.product.deposit.IborFixingDepositCurveNode;
 import com.opengamma.strata.market.product.deposit.TermDepositCurveNode;
-import com.opengamma.strata.market.product.fra.FraCurveNode;
+import com.opengamma.strata.market.product.index.IborFutureCurveNode;
 import com.opengamma.strata.market.product.swap.FixedIborSwapCurveNode;
 import com.opengamma.strata.market.product.swap.FixedOvernightSwapCurveNode;
-import com.opengamma.strata.market.product.swap.IborIborSwapCurveNode;
 import com.opengamma.strata.market.sensitivity.PointSensitivityBuilder;
 import com.opengamma.strata.pricer.deposit.DiscountingIborFixingDepositProductPricer;
 import com.opengamma.strata.pricer.deposit.DiscountingTermDepositProductPricer;
-import com.opengamma.strata.pricer.fra.DiscountingFraTradePricer;
-import com.opengamma.strata.pricer.rate.ImmutableRatesProvider;
+import com.opengamma.strata.pricer.index.DiscountingIborFutureTradePricer;
 import com.opengamma.strata.pricer.rate.RatesProvider;
 import com.opengamma.strata.pricer.sensitivity.MarketQuoteSensitivityCalculator;
 import com.opengamma.strata.pricer.swap.DiscountingSwapProductPricer;
@@ -82,29 +78,29 @@ import com.opengamma.strata.product.deposit.type.IborFixingDepositTemplate;
 import com.opengamma.strata.product.deposit.type.ImmutableTermDepositConvention;
 import com.opengamma.strata.product.deposit.type.TermDepositConvention;
 import com.opengamma.strata.product.deposit.type.TermDepositTemplate;
-import com.opengamma.strata.product.fra.ResolvedFraTrade;
-import com.opengamma.strata.product.fra.type.FraTemplate;
+import com.opengamma.strata.product.index.ResolvedIborFutureTrade;
+import com.opengamma.strata.product.index.type.IborFutureConvention;
+import com.opengamma.strata.product.index.type.IborFutureTemplate;
+import com.opengamma.strata.product.index.type.ImmutableIborFutureConvention;
 import com.opengamma.strata.product.swap.ResolvedSwap;
 import com.opengamma.strata.product.swap.ResolvedSwapTrade;
 import com.opengamma.strata.product.swap.SwapTrade;
+import com.opengamma.strata.product.swap.type.FixedIborSwapConventions;
 import com.opengamma.strata.product.swap.type.FixedIborSwapTemplate;
 import com.opengamma.strata.product.swap.type.FixedOvernightSwapTemplate;
-import com.opengamma.strata.product.swap.type.IborIborSwapConventions;
-import com.opengamma.strata.product.swap.type.IborIborSwapTemplate;
 
 /**
  * Test for curve calibration with 2 curves in USD.
  * One curve is Discounting and Fed Fund forward and the other one is Libor 3M forward.
  */
 @Test
-public class CalibrationZeroRateUsd3OisIrsBsTest {
+public class CalibrationZeroRateUsd2OisFuturesIrsTest {
 
   private static final LocalDate VAL_DATE = LocalDate.of(2015, 7, 21);
 
   private static final CurveInterpolator INTERPOLATOR_LINEAR = CurveInterpolators.LINEAR;
   private static final CurveExtrapolator EXTRAPOLATOR_FLAT = CurveExtrapolators.FLAT;
   private static final DayCount CURVE_DC = ACT_365F;
-  private static final LocalDateDoubleTimeSeries TS_EMTPY = LocalDateDoubleTimeSeries.empty();
 
   // reference data
   private static final ReferenceData REF_DATA = ReferenceData.standard();
@@ -116,8 +112,6 @@ public class CalibrationZeroRateUsd3OisIrsBsTest {
   private static final CurveName DSCON_CURVE_NAME = CurveName.of(DSCON_NAME);
   private static final String FWD3_NAME = "USD-LIBOR3M-FRAIRS";
   private static final CurveName FWD3_CURVE_NAME = CurveName.of(FWD3_NAME);
-  private static final String FWD6_NAME = "USD-LIBOR6M-FRABS";
-  private static final CurveName FWD6_CURVE_NAME = CurveName.of(FWD6_NAME);
   /** Curves associations to currencies and indices. */
   private static final Map<CurveName, Currency> DSC_NAMES = new HashMap<>();
   private static final Map<CurveName, Set<Index>> IDX_NAMES = new HashMap<>();
@@ -130,12 +124,6 @@ public class CalibrationZeroRateUsd3OisIrsBsTest {
     Set<Index> usdLibor3Set = new HashSet<>();
     usdLibor3Set.add(USD_LIBOR_3M);
     IDX_NAMES.put(FWD3_CURVE_NAME, usdLibor3Set);
-    Set<Index> usdLibor6Set = new HashSet<>();
-    usdLibor6Set.add(USD_LIBOR_6M);
-    IDX_NAMES.put(FWD6_CURVE_NAME, usdLibor6Set);
-    TS.put(USD_FED_FUND, TS_EMTPY);
-    TS.put(USD_LIBOR_3M, TS_EMTPY);
-    TS.put(USD_LIBOR_6M, TS_EMTPY);
   }
 
   /** Data for USD-DSCON curve */
@@ -180,73 +168,37 @@ public class CalibrationZeroRateUsd3OisIrsBsTest {
   /* Market values */
   private static final double[] FWD3_MARKET_QUOTES = new double[] {
     0.00236600,
-    0.00258250, 0.00296050,
-    0.00294300, 0.00503000, 0.00939150, 0.01380800, 0.01732000,
+    0.9975, 0.9975, 0.9950, 0.9950, 0.9940, 0.9930, 0.9920, 0.9910,
+    0.00939150, 0.01380800, 0.01732000,
     0.02000000, 0.02396200, 0.02500000, 0.02700000, 0.02930000 };
   private static final int FWD3_NB_NODES = FWD3_MARKET_QUOTES.length;
   private static final String[] FWD3_ID_VALUE = new String[] {
     "USD-Fixing-3M",
-    "USD-FRA3Mx6M", "USD-FRA6Mx9M",
-    "USD-IRS3M-1Y", "USD-IRS3M-2Y", "USD-IRS3M-3Y", "USD-IRS3M-4Y", "USD-IRS3M-5Y",
+    "USD-ED1", "USD-ED2", "USD-ED3", "USD-ED4", "USD-ED5", "USD-ED6", "USD-ED7", "USD-ED8", 
+    "USD-IRS3M-3Y", "USD-IRS3M-4Y", "USD-IRS3M-5Y",
     "USD-IRS3M-6Y", "USD-IRS3M-7Y", "USD-IRS3M-8Y", "USD-IRS3M-9Y", "USD-IRS3M-10Y" };
   /* Nodes */
   private static final CurveNode[] FWD3_NODES = new CurveNode[FWD3_NB_NODES];
   /* Tenors */
-  private static final Period[] FWD3_FRA_TENORS = new Period[] { // Period to start
-      Period.ofMonths(3), Period.ofMonths(6)};
-  private static final int FWD3_NB_FRA_NODES = FWD3_FRA_TENORS.length;
+  private static final int[] FWD3_FUT_SEQ = new int[] {1, 2, 3, 4, 5, 6, 7, 8};
+  private static final int FWD3_NB_FUT_NODES = FWD3_FUT_SEQ.length;
   private static final Period[] FWD3_IRS_TENORS = new Period[] {
-    Period.ofYears(1), Period.ofYears(2), Period.ofYears(3), Period.ofYears(4), Period.ofYears(5),
+    Period.ofYears(3), Period.ofYears(4), Period.ofYears(5),
     Period.ofYears(6), Period.ofYears(7), Period.ofYears(8), Period.ofYears(9), Period.ofYears(10)};
   private static final int FWD3_NB_IRS_NODES = FWD3_IRS_TENORS.length;
   static {
     FWD3_NODES[0] = IborFixingDepositCurveNode.of(IborFixingDepositTemplate.of(USD_LIBOR_3M),
         QuoteId.of(StandardId.of(SCHEME, FWD3_ID_VALUE[0])));
-    for (int i = 0; i < FWD3_NB_FRA_NODES; i++) {
-      FWD3_NODES[i + 1] = FraCurveNode.of(FraTemplate.of(FWD3_FRA_TENORS[i], USD_LIBOR_3M),
+    IborFutureConvention convention = ImmutableIborFutureConvention.of(USD_LIBOR_3M, DateSequences.QUARTERLY_IMM);
+    for (int i = 0; i < FWD3_NB_FUT_NODES; i++) {
+      IborFutureTemplate template = IborFutureTemplate.of(Period.ofDays(7), FWD3_FUT_SEQ[i], convention);
+      FWD3_NODES[i + 1] = IborFutureCurveNode.of(template, 
           QuoteId.of(StandardId.of(SCHEME, FWD3_ID_VALUE[i + 1])));
     }
     for (int i = 0; i < FWD3_NB_IRS_NODES; i++) {
-      FWD3_NODES[i + 1 + FWD3_NB_FRA_NODES] = FixedIborSwapCurveNode.of(
+      FWD3_NODES[i + 1 + FWD3_NB_FUT_NODES] = FixedIborSwapCurveNode.of(
           FixedIborSwapTemplate.of(Period.ZERO, Tenor.of(FWD3_IRS_TENORS[i]), USD_FIXED_6M_LIBOR_3M),
-          QuoteId.of(StandardId.of(SCHEME, FWD3_ID_VALUE[i + 1 + FWD3_NB_FRA_NODES])));
-    }
-  }
-
-  /** Data for USD-LIBOR6M curve */
-  /* Market values */
-  private static final double[] FWD6_MARKET_QUOTES = new double[] {
-      0.00246600,
-      0.00268250,
-      0.00130000, 0.00120000, 0.00100000, 0.00100000, 0.00100000, 
-      0.00100000, 0.00100000, 0.00090000, 0.00090000, 0.00080000, };
-  private static final int FWD6_NB_NODES = FWD6_MARKET_QUOTES.length;
-  private static final String[] FWD6_ID_VALUE = new String[] {
-      "Fixing",
-      "FRA3Mx9M", 
-      "USD-BS36-1Y", "USD-BS36-2Y", "USD-BS36-3Y", "USD-BS36-4Y", "USD-BS36-5Y",
-      "USD-BS36-6Y", "USD-BS36-7Y", "USD-BS36-8Y", "USD-BS36-9Y", "USD-BS36-10Y"};
-  /* Nodes */
-  private static final CurveNode[] FWD6_NODES = new CurveNode[FWD6_NB_NODES];
-  /* Tenors */
-  private static final Period[] FWD6_FRA_TENORS = new Period[] { // Period to start
-      Period.ofMonths(3)};
-  private static final int FWD6_NB_FRA_NODES = FWD6_FRA_TENORS.length;
-  private static final Period[] FWD6_IRS_TENORS = new Period[] {
-    Period.ofYears(1), Period.ofYears(2), Period.ofYears(3), Period.ofYears(4), Period.ofYears(5),
-    Period.ofYears(6), Period.ofYears(7), Period.ofYears(8), Period.ofYears(9), Period.ofYears(10)};
-  private static final int FWD6_NB_IRS_NODES = FWD6_IRS_TENORS.length;
-  static {
-    FWD6_NODES[0] = IborFixingDepositCurveNode.of(IborFixingDepositTemplate.of(USD_LIBOR_6M),
-        QuoteId.of(StandardId.of(SCHEME, FWD6_ID_VALUE[0])));
-    for (int i = 0; i < FWD6_NB_FRA_NODES; i++) {
-      FWD6_NODES[i + 1] = FraCurveNode.of(FraTemplate.of(FWD6_FRA_TENORS[i], USD_LIBOR_6M),
-          QuoteId.of(StandardId.of(SCHEME, FWD6_ID_VALUE[i + 1])));
-    }
-    for (int i = 0; i < FWD6_NB_IRS_NODES; i++) {
-      FWD6_NODES[i + 1 + FWD6_NB_FRA_NODES] = IborIborSwapCurveNode.of(
-          IborIborSwapTemplate.of(Period.ZERO, Tenor.of(FWD6_IRS_TENORS[i]), USD_LIBOR_3M_LIBOR_6M),
-          QuoteId.of(StandardId.of(SCHEME, FWD6_ID_VALUE[i + 1 + FWD6_NB_FRA_NODES])));
+          QuoteId.of(StandardId.of(SCHEME, FWD3_ID_VALUE[i + 1 + FWD3_NB_FUT_NODES])));
     }
   }
 
@@ -260,9 +212,6 @@ public class CalibrationZeroRateUsd3OisIrsBsTest {
     for (int i = 0; i < FWD3_NB_NODES; i++) {
       builder.addValue(QuoteId.of(StandardId.of(SCHEME, FWD3_ID_VALUE[i])), FWD3_MARKET_QUOTES[i]);
     }
-    for (int i = 0; i < FWD6_NB_NODES; i++) {
-      builder.addValue(QuoteId.of(StandardId.of(SCHEME, FWD6_ID_VALUE[i])), FWD6_MARKET_QUOTES[i]);
-    }
     ALL_QUOTES = builder.build();
   }
 
@@ -275,9 +224,6 @@ public class CalibrationZeroRateUsd3OisIrsBsTest {
     List<CurveNode[]> groupFwd3 = new ArrayList<>();
     groupFwd3.add(FWD3_NODES);
     CURVES_NODES.add(groupFwd3);
-    List<CurveNode[]> groupFwd6 = new ArrayList<>();
-    groupFwd6.add(FWD6_NODES);
-    CURVES_NODES.add(groupFwd6);
   }
 
   /** All metadata by groups */
@@ -291,16 +237,12 @@ public class CalibrationZeroRateUsd3OisIrsBsTest {
     groupFwd3.add(DefaultCurveMetadata.builder().curveName(FWD3_CURVE_NAME).xValueType(ValueType.YEAR_FRACTION)
         .yValueType(ValueType.ZERO_RATE).dayCount(CURVE_DC).build());
     CURVES_METADATA.add(groupFwd3);
-    List<CurveMetadata> groupFwd6 = new ArrayList<>();
-    groupFwd6.add(DefaultCurveMetadata.builder().curveName(FWD6_CURVE_NAME).xValueType(ValueType.YEAR_FRACTION)
-        .yValueType(ValueType.ZERO_RATE).dayCount(CURVE_DC).build());
-    CURVES_METADATA.add(groupFwd6);
   }
 
   private static final DiscountingIborFixingDepositProductPricer FIXING_PRICER =
       DiscountingIborFixingDepositProductPricer.DEFAULT;
-  private static final DiscountingFraTradePricer FRA_PRICER =
-      DiscountingFraTradePricer.DEFAULT;
+  private static final DiscountingIborFutureTradePricer FUT_PRICER =
+      DiscountingIborFutureTradePricer.DEFAULT;
   private static final DiscountingSwapProductPricer SWAP_PRICER =
       DiscountingSwapProductPricer.DEFAULT;
   private static final DiscountingTermDepositProductPricer DEPO_PRICER =
@@ -334,49 +276,16 @@ public class CalibrationZeroRateUsd3OisIrsBsTest {
           .extrapolatorLeft(EXTRAPOLATOR_FLAT)
           .extrapolatorRight(EXTRAPOLATOR_FLAT)
           .nodes(FWD3_NODES).build();
-  private static final InterpolatedNodalCurveDefinition FWD6_CURVE_DEFN =
-      InterpolatedNodalCurveDefinition.builder()
-          .name(FWD6_CURVE_NAME)
-          .xValueType(ValueType.YEAR_FRACTION)
-          .yValueType(ValueType.ZERO_RATE)
-          .dayCount(CURVE_DC)
-          .interpolator(INTERPOLATOR_LINEAR)
-          .extrapolatorLeft(EXTRAPOLATOR_FLAT)
-          .extrapolatorRight(EXTRAPOLATOR_FLAT)
-          .nodes(FWD6_NODES).build();
   private static final CurveGroupDefinition CURVE_GROUP_CONFIG =
       CurveGroupDefinition.builder()
           .name(CURVE_GROUP_NAME)
           .addCurve(DSC_CURVE_DEFN, USD, USD_FED_FUND)
-          .addForwardCurve(FWD3_CURVE_DEFN, USD_LIBOR_3M)
-          .addForwardCurve(FWD6_CURVE_DEFN, USD_LIBOR_6M).build();
-
-  private static final CurveGroupDefinition GROUP_1 =
-      CurveGroupDefinition.builder()
-          .name(CurveGroupName.of(DSCON_NAME))
-          .addCurve(DSC_CURVE_DEFN, USD, USD_FED_FUND)
-          .build();
-  private static final CurveGroupDefinition GROUP_2 =
-      CurveGroupDefinition.builder()
-          .name(CurveGroupName.of(FWD3_NAME))
-          .addForwardCurve(FWD3_CURVE_DEFN, USD_LIBOR_3M)
-          .build();
-  private static final CurveGroupDefinition GROUP_3 =
-      CurveGroupDefinition.builder()
-          .name(CurveGroupName.of(FWD6_NAME))
-          .addForwardCurve(FWD6_CURVE_DEFN, USD_LIBOR_6M)
-          .build();
-  private static final ImmutableRatesProvider KNOWN_DATA = ImmutableRatesProvider.builder(VAL_DATE).build();
+          .addForwardCurve(FWD3_CURVE_DEFN, USD_LIBOR_3M).build();
 
   //-------------------------------------------------------------------------
   public void calibration_present_value_oneGroup() {
-    RatesProvider result = CALIBRATOR.calibrate(CURVE_GROUP_CONFIG, VAL_DATE, ALL_QUOTES, REF_DATA, TS);
-    assertPresentValue(result);
-  }
-
-  public void calibration_present_value_threeGroups() {
     RatesProvider result =
-        CALIBRATOR.calibrate(ImmutableList.of(GROUP_1, GROUP_2, GROUP_3), KNOWN_DATA, ALL_QUOTES, REF_DATA);
+        CALIBRATOR.calibrate(CURVE_GROUP_CONFIG, VAL_DATE, ALL_QUOTES, REF_DATA, TS);
     assertPresentValue(result);
   }
   
@@ -386,21 +295,14 @@ public class CalibrationZeroRateUsd3OisIrsBsTest {
         marketData -> CALIBRATOR.calibrate(CURVE_GROUP_CONFIG, VAL_DATE, marketData, REF_DATA, TS);
     calibration_market_quote_sensitivity_check(f, shift);
   }
-  
-  public void calibration_market_quote_sensitivity_three_group() {
-    double shift = 1.0E-6;
-    Function<MarketData, RatesProvider> calibrator =
-        marketData -> CALIBRATOR.calibrate(ImmutableList.of(GROUP_1, GROUP_2, GROUP_3), KNOWN_DATA, marketData, REF_DATA);
-    calibration_market_quote_sensitivity_check(calibrator, shift);
-  }
 
   private void calibration_market_quote_sensitivity_check(
       Function<MarketData, RatesProvider> calibrator,
       double shift) {
     double notional = 100_000_000.0;
     double spread = 0.0050;
-    SwapTrade trade = IborIborSwapConventions.USD_LIBOR_3M_LIBOR_6M.createTrade(
-        VAL_DATE, Period.ofMonths(8), Tenor.TENOR_7Y, BuySell.BUY, notional, spread, REF_DATA);
+    SwapTrade trade = FixedIborSwapConventions.USD_FIXED_1Y_LIBOR_3M
+        .createTrade(VAL_DATE, Period.ofMonths(8), Tenor.TENOR_7Y, BuySell.BUY, notional, spread, REF_DATA);
     RatesProvider result = calibrator.apply(ALL_QUOTES);
     ResolvedSwap product = trade.getProduct().resolve(REF_DATA);
     PointSensitivityBuilder pts = SWAP_PRICER.presentValueSensitivity(product, result);
@@ -424,15 +326,6 @@ public class CalibrationZeroRateUsd3OisIrsBsTest {
       RatesProvider rpShifted = calibrator.apply(marketData);
       double pvS = SWAP_PRICER.presentValue(product, rpShifted).getAmount(USD).getAmount();
       assertEquals(mqsFwd3Computed[i], (pvS - pv0) / shift, TOLERANCE_PV_DELTA, "FWD3 - node " + i);
-    }
-    double[] mqsFwd6Computed = mqs.getSensitivity(FWD6_CURVE_NAME, USD).getSensitivity().toArray();
-    for (int i = 0; i < FWD6_NB_NODES; i++) {
-      Map<MarketDataId<?>, Object> map = new HashMap<>(ALL_QUOTES.getValues());
-      map.put(QuoteId.of(StandardId.of(SCHEME, FWD6_ID_VALUE[i])), FWD6_MARKET_QUOTES[i] + shift);
-      ImmutableMarketData marketData = ImmutableMarketData.of(VAL_DATE, map);
-      RatesProvider rpShifted = calibrator.apply(marketData);
-      double pvS = SWAP_PRICER.presentValue(product, rpShifted).getAmount(USD).getAmount();
-      assertEquals(mqsFwd6Computed[i], (pvS - pv0) / shift, TOLERANCE_PV_DELTA, "FWD6 - node " + i);
     }
   }
 
@@ -465,38 +358,16 @@ public class CalibrationZeroRateUsd3OisIrsBsTest {
     CurrencyAmount pvFixing3 = FIXING_PRICER.presentValue(
         ((ResolvedIborFixingDepositTrade) fwd3Trades.get(0)).getProduct(), result);
     assertEquals(pvFixing3.getAmount(), 0.0, TOLERANCE_PV);
-    // FRA
-    for (int i = 0; i < FWD3_NB_FRA_NODES; i++) {
-      CurrencyAmount pvFra = FRA_PRICER.presentValue(
-          ((ResolvedFraTrade) fwd3Trades.get(i + 1)), result);
-      assertEquals(pvFra.getAmount(), 0.0, TOLERANCE_PV);
+    // Futures
+    for (int i = 0; i < FWD3_NB_FUT_NODES; i++) {
+      CurrencyAmount pvFut = FUT_PRICER.presentValue(
+          ((ResolvedIborFutureTrade) fwd3Trades.get(i + 1)), result, 0.0);
+      assertEquals(pvFut.getAmount(), 0.0, TOLERANCE_PV);
     }
     // IRS
     for (int i = 0; i < FWD3_NB_IRS_NODES; i++) {
       MultiCurrencyAmount pvIrs = SWAP_PRICER.presentValue(
-          ((ResolvedSwapTrade) fwd3Trades.get(i + 1 + FWD3_NB_FRA_NODES)).getProduct(), result);
-      assertEquals(pvIrs.getAmount(USD).getAmount(), 0.0, TOLERANCE_PV);
-    }
-    // Test PV Fwd3
-    CurveNode[] fwd6Nodes = CURVES_NODES.get(2).get(0);
-    List<ResolvedTrade> fwd6Trades = new ArrayList<>();
-    for (int i = 0; i < fwd6Nodes.length; i++) {
-      fwd6Trades.add(fwd6Nodes[i].resolvedTrade(VAL_DATE, ALL_QUOTES, REF_DATA));
-    }
-    // Fixing 
-    CurrencyAmount pvFixing6 = FIXING_PRICER.presentValue(
-        ((ResolvedIborFixingDepositTrade) fwd6Trades.get(0)).getProduct(), result);
-    assertEquals(pvFixing6.getAmount(), 0.0, TOLERANCE_PV);
-    // FRA
-    for (int i = 0; i < FWD6_NB_FRA_NODES; i++) {
-      CurrencyAmount pvFra = FRA_PRICER.presentValue(
-          ((ResolvedFraTrade) fwd6Trades.get(i + 1)), result);
-      assertEquals(pvFra.getAmount(), 0.0, TOLERANCE_PV);
-    }
-    // BS
-    for (int i = 0; i < FWD6_NB_IRS_NODES; i++) {
-      MultiCurrencyAmount pvIrs = SWAP_PRICER.presentValue(
-          ((ResolvedSwapTrade) fwd6Trades.get(i + 1 + FWD6_NB_FRA_NODES)).getProduct(), result);
+          ((ResolvedSwapTrade) fwd3Trades.get(i + 1 + FWD3_NB_FUT_NODES)).getProduct(), result);
       assertEquals(pvIrs.getAmount(USD).getAmount(), 0.0, TOLERANCE_PV);
     }
   }
@@ -517,11 +388,11 @@ public class CalibrationZeroRateUsd3OisIrsBsTest {
         count += result.getValuationDate().getDayOfMonth();
       }
       endTime = System.currentTimeMillis();
-      System.out.println("Performance: " + nbTests + " calibrations for 3 curve with 43 nodes in "
+      System.out.println("Performance: " + nbTests + " calibrations for 2 curves with 35 nodes in "
           + (endTime - startTime) + " ms.");
     }
     System.out.println("Avoiding hotspot: " + count);
-    // Previous run: 2325 ms for 100 calibrations (3 curve simultaneous - 43 nodes)
+    // Previous run: 1050 ms for 100 calibrations (2 curves simultaneous - 35 nodes)
   }
 
 }
