+++ modules/pricer/src/test/java/com/opengamma/strata/pricer/curve/CurveCalibratorTest.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.pricer.calibration;
+package com.opengamma.strata.pricer.curve;
 
 import static org.assertj.core.api.Assertions.assertThat;
